Davemills/v5.1 rc1

+++ bin/runtime-generator.tcl
@@ -18,7 +18,7 @@ foreach subsys $EVERYTHING {
 if { $argv == "" || [lsearch $argv idl] > -1 } {
  puts stdout  "Updating IDL only"
  foreach subsys $EVERYTHING {
-  if { [info exists DO($subsys)] } {
+  if { [info exists DO($subsys)] && [info exists SYSDIC($subsys,idl)] } {
    set bad ""
    set result ""
    catch { set results [exec salgenerator $subsys sal idl] } bad
@@ -46,7 +46,7 @@ if { $argv == "" || [lsearch $argv validate] > -1} {
 if { $argv == "" || [lsearch $argv cpp] > -1 } {
  puts stdout  "Generating C++"
  foreach subsys $EVERYTHING {
-  if { [info exists DO($subsys)] } {
+  if { [info exists DO($subsys)] && [info exists SYSDIC($subsys,cpp)] } {
    set bad ""
    set result ""
    catch { set results [exec salgenerator $subsys sal cpp ] } bad
@@ -58,7 +58,7 @@ if { $argv == "" || [lsearch $argv cpp] > -1 } {
 if { $argv == "" || [lsearch $argv python] > -1 } {
  puts stdout  "Generating Python"
  foreach subsys $EVERYTHING {
-  if { [info exists DO($subsys)] } {
+  if { [info exists DO($subsys)] && [info exists SYSDIC($subsys,salpy)] } {
    set bad ""
    set result ""
    catch { set results [exec salgenerator $subsys sal python ] } bad
@@ -71,23 +71,10 @@ if { $argv == "" || [lsearch $argv python] > -1 } {
 if { $argv == "" || [lsearch $argv labview] > -1 } {
  puts stdout  "Generating LabVIEW"
  foreach subsys $EVERYTHING {
-  if { [info exists DO($subsys)] } {
+  if { [info exists DO($subsys)] && [info exists SYSDIC($subsys,labview)] } {
    set bad ""
    set result ""
    catch { set results [exec salgenerator $subsys labview ] } bad
-   puts stdout "$result $bad"
-  }
- }
-}
-
-
-if { $argv == "" || [lsearch $argv lib] > -1 } {
- puts stdout  "Updating libraries"
- foreach subsys $EVERYTHING {
-  if { [info exists DO($subsys)] } {
-   set bad ""
-   set result ""
-   catch { set results [exec salgenerator $subsys lib ] } bad
    puts stdout "$result $bad"
   }
  }
@@ -99,7 +86,7 @@ if { $argv == "" || [lsearch $argv java] > -1 } {
  puts stdout  "Generating Java"
  foreach subsys $EVERYTHING {
   if { $subsys != "MTMount" } {
-   if { [info exists DO($subsys)] } {
+   if { [info exists DO($subsys)] && [info exists SYSDIC($subsys,java)] } {
     exec rm -fr [set subsys]/java
     set bad ""
     set result ""
@@ -113,6 +100,19 @@ if { $argv == "" || [lsearch $argv java] > -1 } {
   }
  }
 }
+
+if { $argv == "" || [lsearch $argv lib] > -1 } {
+ puts stdout  "Updating libraries"
+ foreach subsys $EVERYTHING {
+  if { [info exists DO($subsys)] } {
+   set bad ""
+   set result ""
+   catch { set results [exec salgenerator $subsys lib ] } bad
+   puts stdout "$result $bad"
+  }
+ }
+}
+
 
 
 cd $env(SAL_WORK_DIR)
